rag 1/requirements.txt
Django==4.2.24
python-dotenv==1.0.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
transformers==4.43.3
accelerate==0.33.0
numpy==1.26.4
pandas==2.2.2
pypdf==4.3.1
python-docx==1.1.2
google-generativeai==0.7.2
groq==0.9.0
whitenoise==6.7.0
django-cors-headers==4.4.0
